Deleted nested scroll listview on expandableList_section. Create IconItem model and put the static data into Repo. Removed ~ItemSection files
