--- lib/widgets/iconlistitem_section.dart
@@ -1,29 +1,0 @@
-
-import 'package:flutter/material.dart';
-
-import '../models/listitem.dart';
-
-class IconListItem implements ListItem {
-
-  String name;
-  String latestAccess;
-  IconData icon;
-  IconListItem(this.name, this.latestAccess, this.icon);
-
-
-  @override
-  Widget buildBody(BuildContext context) {
-
-    return   ListTile(
-      visualDensity: VisualDensity(horizontal: 0, vertical: -3),
-          title: Text(name),
-          subtitle: Text(latestAccess),
-          leading: CircleAvatar(
-            child: Icon(icon),
-          ),
-    );
-
-
-  }
-
-}
